Source/requirements.txt
customtkinter
pynput
packaging
darkdetect
pillow
